requirements.txt
undetected-chromedriver==3.5.5
selenium==4.22.0
PySocks==1.7.1
setuptools==70.1.1
beautifulsoup4==4.12.3
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.0
websockets==12.0
redis==5.0.7
celery==5.4.0
python-socketio==5.11.3
python-telegram-bot==21.4
aioredis==2.0.1
aiohttp==3.9.5
translatepy==2.3
fastapi==0.112.1
aiofiles==24.1.0
flower==2.0.1
alembic==1.14.0
sqlalchemy==2.0.36
asyncpg==0.30.0